WIP 5

+++ includes/libft.h
@@ -37,6 +37,6 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void	ft_putchar_fd(char c, int fd);
 void	ft_putstr_fd(char *s, int fd);
 void	ft_putendl_fd(char *s, int fd);
-void	ft_putnbr_fd(unsigned int n, int fd);
+void	ft_putnbr_fd(long long n, int fd);
 
 #endif
